优化分页列表的处理逻辑

+++ mate-platform/mate-system/src/main/java/vip/mate/system/controller/SysLogController.java
@@ -1,20 +1,19 @@
 package vip.mate.system.controller;
 
 
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
-
 import vip.mate.core.auth.annotation.EnableToken;
 import vip.mate.core.common.api.Result;
+import vip.mate.core.database.entity.Search;
 import vip.mate.core.log.annotation.Log;
 import vip.mate.core.web.controller.BaseController;
 import vip.mate.core.web.util.CollectionUtil;
 import vip.mate.system.service.ISysLogService;
-
-import java.util.Map;
 
 /**
  * <p>
@@ -42,8 +41,8 @@ public class SysLogController extends BaseController {
             @ApiImplicitParam(name = "startDate", required = true, value = "创建开始日期", paramType = "form"),
             @ApiImplicitParam(name = "endDate", required = true, value = "创建结束日期", paramType = "form"),
     })
-    public Result<?> list(@RequestParam Map<String, String> query) {
-        return Result.data(sysLogService.listPage(query));
+    public Result<?> list(Page page, Search search) {
+        return Result.data(sysLogService.listPage(page, search));
     }
 
     @EnableToken
@@ -54,6 +53,9 @@ public class SysLogController extends BaseController {
             @ApiImplicitParam(name = "ids", required = true, value = "多个用,号隔开", paramType = "form")
     })
     public Result<?> delete(@RequestParam String ids) {
+        if (true) {
+            return Result.fail("演示环境，不能删除日志");
+        }
         if (sysLogService.removeByIds(CollectionUtil.stringToCollection(ids))){
             return Result.success("删除成功");
         }
